[REMOVE] redundant cast

+++ src/main/java/com/github/unchama/seichiassist/listener/PlayerBlockBreakListener.java
@@ -141,7 +141,7 @@ public class PlayerBlockBreakListener implements Listener {
 		if(!playerdata.getActiveskilldata().skillcanbreakflag){
 			//SEを再生
 			if(SeichiAssist.DEBUG) player.sendMessage(ChatColor.RED + "クールタイムの破壊");
-			player.playSound(player.getLocation(), Sound.BLOCK_DISPENSER_FAIL, (float)0.5, 1);
+			player.playSound(player.getLocation(), Sound.BLOCK_DISPENSER_FAIL, 0.5f, 1);
 			return;
 		}
